Changed the JSON between compiled and edited.
The edited JSON is making more readable by just listing the properties
of objects and actions (it is more easy to manually edit now, because
the execution functions are not added).
The generated JSON had been lightened by just putting the pertinent
information and the execution functions for the actions.