Added a reload feature and bug fixes.
Fixed a bug that I introduced in the previous commit, with the preview.
Fixed a bug on windows (the preview didn't run, now it does).
Added a reload feature for a project (when modifications are made
manually on the dahu file, the project can be reloaded without
specifying again the project directory, to build it with the manual
modifications).