generate presentation in index.html, not presentation.html
The fact that the file is an HTML file is an implementation detail, and
should not be visible in the URL (who knows, we may need an index.php or
index.shtml in the future). By naming the file index.html, the public URL
of the presentation becomes the directory name, which is shorter,
simpler, and stable.