Added the possibility to change the capture key.
The capture key can be changed via the graphical interface.
A file named 'config.json' contains this key.
This file is stored in a directory named '.dahuapp' which is
created in the user home directory.

Currently the only keys available as capture keys are the
function keys (F*).

There is also a minor fix for Windows (when an empty string
was typed in the opening/creation of project, an error occured).