Anchors available for the final presentation.
For the final presentation in the web browser, an anchor can be specified.
As in the final presentation, the slides are "forgotten" and only the
actions can be referenced, this anchor specifies the starting action (see
the user's manual update for more details).

For previous versions of Dahu projects, an automatic procedure will add
the missing information to the project file to enable the anchors.