Platform-compatibility improved
On some linux platforms (for example on my Ubuntu 13.04 32 bits) the
java.awt.Desktop API is not supported and causes that the application
crashes instantaneously when it is run.
The java.awt.Desktop class is used to have access to the default web
browser of the platform.

To avoid this problem, I added
a java library (BrowserLauncher2) which is an opensource API for Java
made for web browser opening. As it is an external library, it is only
used when java.awt.Desktop is not supported, so the Desktop API has
the priority when it works.