autoPlay and autoStart parameters
These parameters allow running a presentation non-interactively.
'autoStart' controls when the presentation starts, and 'autoPlay'
whether to pause when reaching 'onClick' actions.

They are encoded as GET parameters to avoid interfering with the
anchor part. It's still possible to use <a href="#something">
to navigate within the presentation. One downside of this is that
changing one auto* parameter requires reloading the page.