Work around "jar:" issue with local files
When running in bundle mode (i.e. as .jar file), we used to load
dahuapp.html and the associated CSS and HTML files using a "jar:file:"
URL. This hits the bug https://javafx-jira.kenai.com/browse/RT-33570 and
prevents loading local images.

To work around the issue, we load dahuapp.html as a string, using
loadContent(). Relative URLs use localresource: in the HTML source, and
are substituted to the actual jar: resource URL before being loaded.
Hacky, but works :-\.

For some reasons, local file access is allowed with "file:/path/to/file",
but not with "/path/to/file" (I guess the webview doen't have a default
protocol since the page is loaded without URL).