remove global Handlebars helpers
Previously the Screencast Controller was defining global Handlebars helpers (namely: screencastWidth, screencastHeight).
However those helpers are only used in the generation process by the compiler module.
Besides the fact that defining global helpers is not elegant, the main drawback of this approach was that it adds a useless dependency between the Screencast Controller and the compiler module. Such dependency will be a problem when trying to compile a screencast from the CLI since the CLI will not be able to instantiate Marionette.Controllers (Marionette is not a dependency that can be used server side or in Nashorn due to some bad dependencies to Backbone.$ like Backbone.$.deferred).