Addressing issue #113.
Getting rid of "edu.stanford.ejalbert.BrowserLauncher", it wasn't
working as we need anyway (bad thread handling).

Currently, if the "java.awt.Desktop" utilities are not available,
we can't open the preview directly from the dahu app.

The app closes normally now (it was mainly caused by the
bad handling of the ExecutorService termination in the Browser module).