Fixes issue #138
This fixes the fact that the default browser didn't open when selecting
"Preview" on the menu. A workaround is used for non-GNOME systems
where xdg-open is activated to manage default applications : tested and
verified to work on Debian Squeeze with KDE.
A verification for Linux system is added to make sure it is possible to
run xdg-open and get the default browser.