Introduce support for note edition and preview
Insert the possibility of adding notes
to the presentation. In the editor, a textarea
can be filled with a text for each screen of the
current screencast. Those notes can be read in
the preview with a pluggin that can display/hide.

Edit and preview notes :

After taking a screenshot, the creator can select the screenshot to which he
wants to add a note by clicking on it. When the screen is selected, the
workspace displays the screenshot and under it there is a note area in which the
creator can write the note concerning the selected screenshot. When ths note
area loses the focus, its content is saved in the model but note in the ".dahu"
file. To save his notes, the user must click on the "save" menu item in the
"file" menu in the toolbar. At that moment the note is saved in the ".dahu" file
 and when the presentation is generated, the notes saved will present in the
presentation.

During a presentation, to display a note, the user can display a note by
pressing the key "n" when he is on the slide which contains a note. Warning, the
notes are displayed only if the user uses chrome (no support on mozilla).

Added :
- a 'note model'
- a 'note layout'
- a 'note view'

Note model only contains a single text and an Id (for
the DOM). Note View catches events related to changing
screen and saving, in order to load and save the notes.
Currently the note is saved on the textarea 'focusout'
and on the action save in the menu (wich doesn't
trigger 'focusout').

Modification :

- bower.json : add dependencies to make the
pluggin note work with deckJS
- Gruntfile.js : allow the copy of deck.js
files to the right location
- Screen Model : include a 'note' attribute
- Compiler et template Screencast : change
the access to the datas given by the model with
a .toJSON() function.
- Dahuapp
- Layout Workspace : add a note region to the
layout.