Adressing issue #133
Implementing the actions when clicking on the buttons of the webview.
The fullscreen button calls the editor setMaximized to true or false,
depending on its previous value.
The preview button in the interface has the same action as the button in the
editor (previewScreencast).
The configuration menu for the screencast is not implemented yet : throws error