Allow to sort collections using drag and drop
File added : sortable behavior
Define the behavior "sortable" that is applied on the childViewContainer of
the view on which is added the behavior "sortable" if the
childViewContainer is defined or it is applied to the element itself otherwise.

File Modified : Actions' list view and screens' list view
Add the behavior "sortable" on the view of the actions and the screens.